+++ src/hyperactive/optimization/gradient_free_optimizers/_composite_optimizer.py
