+++ src/hyperactive/optimization/gradient_free_optimizers/_constraint.py
