+++ src/hyperactive/optimization/gradient_free_optimizers/_optimizer_attributes.py
