+++ src/hyperactive/optimization/gradient_free_optimizers/_optimizers.py
@@ -3,7 +3,7 @@
 # License: MIT License
 
 
-from .._optimizer_api import BaseOptimizer
+from ._base_optimizer import BaseOptimizer
 
 from gradient_free_optimizers import (
     HillClimbingOptimizer as _HillClimbingOptimizer,
