+++ src/hyperactive/optimization/gradient_free_optimizers/_print_results.py
