+++ src/hyperactive/optimization/gradient_free_optimizers/_run_info.py
