+++ src/hyperactive/optimization/gradient_free_optimizers/_search_info.py
