+++ src/hyperactive/optimization/gradient_free_optimizers/search_space.py
