+++ src/hyperactive/optimization/talos/__init__.py
@@ -1,0 +1,1 @@
+from ._optimizers import TalosOptimizer
