+++ src/hyperactive/optimization/talos/_optimizers.py
@@ -1,0 +1,12 @@
+from .._abstract_optimizer import AbstractOptimizer
+
+
+class TalosOptimizer(AbstractOptimizer):
+    def __init__(self):
+        pass
+
+    def add_search(self):
+        pass
+
+    def run(self):
+        pass
